The current approach for running the E2E tests through `kubetest` is to let the tool magically download kubernetes and start the cluster. (#86)
To ensure that our tests don't unexpectedly break, we don't pin the kubernetes version to a particular one (https://github.com/knative/serving/pull/1294) but always use `latest`.

However, the concept of `latest` for kubetests means "kubernetes head", which can be broken (didn't happen so far) or doesn't support all platforms (OS X is the most common case).

Unfortunately, switching to `default` or `release/stable` translates to kubernetes 1.9, which is not recommended for Knative according to the docs.

This PR adds the function `download_k8s()`: it will download the latest public, stable GKE binary supported by the test cluster and the client machine.

As a bonus, the function knocks out the `kubernetes-test.tar.gz` package; this 1.2GB package is not used by Knative tests, and is a severe burden on slower connections (anedoctal evidence: 20+ min download on my home network).

Fixes #40.

Bonus: fixes #80.