precompute start and finish time of prow builds (#435)
Start time is needed for sorting gcs builds by start time, which is obtained by parsing  in gcs bucket. Current implementation queries gcs for every comparison, which is very time consuming and error prone. This PR addresses this problem by precompute start time so that the performance is much better and more reliable