fix(dep-collector): Rebuild dep-collector unconditionally
Don't reuse an existing dep-collector in /go/bin as this might be broken
and might be created by a different job which requires a different
version.
Instead rebuild dep-collector for each run afresh, either by
picking it up from a vendored dependency,
or, if toosl/dep-collector is not vendored in the project under test,
then install the latest version directly over the air.

The `go get` for licenseclassifier has been removed as this will be
performed implicitely by the build of dep-collector anyway.

I would recommend to switch over for other go tools, too, to *not*
pick up existing binaries especially when there are different versions
used in different projects (which is why there is support for installing
from the vendor dir, right ?). So `run_go_tool` looks broken to me anyway.