[coverage] update file path converter
The change of repo structure from github.com/$REPO_OWNER/$REPO_NAME to knative.dev/$REPO_NAME/ breaks the function FilePathProfileToGithub, because the function is based on the repo structure.
This change introduced a more adaptive way to make the path conversion. It no longer relies on how the where and how the repo is hosted. Instead it uses repository root and GOPATH to determine exact prefix to remove to make the conversion.