[coverage] update file path converter (#1290)
* [coverage] update string mapper to accommodate repo structure change

The change of repo structure from github.com/$REPO_OWNER/$REPO_NAME to knative.dev/$REPO_NAME/ breaks the function FilePathProfileToGithub, because the function is based on the repo structure. This change is to fix the issue.

* [coverage] update file path converter

The change of repo structure from github.com/$REPO_OWNER/$REPO_NAME to knative.dev/$REPO_NAME/ breaks the function FilePathProfileToGithub, because the function is based on the repo structure.
This change introduced a more adaptive way to make the path conversion. It no longer relies on how the where and how the repo is hosted. Instead it uses repository root and GOPATH to determine exact prefix to remove to make the conversion.

* nit

* nit 2

* added test case for additional repo structure

* address Joyce's comment

* changed test cases to tabular form and mocked both scenarios

* reset getRepoRoot to unmocked version after test

* fix

* add doc string

* improve defer func