Automatically generate the Prow plugins config (#1465)
* Automatically generate the Prow plugins config

Repos with plugins enabled are those with tide enabled.

The autogenerated plugins config file will be longer now because each repo is specified explicitly.

Currently all repos have the same set of plugins enabled.

There are no changes in which plugins are enabled for any repos, compared to the previous static config.

* Add comments for the plugins config.