Improve Makefile usage and document existing
Makefiles should not be magically included by other Makefiles, so
moving common functionality into `prod_staging.mk` file and
documenting it.

The config generator Go program was giving old information about
how to generate files, so corrected it.

The generated configs now depend on the templates for the config
generator, so they can be auto-updated as necessary.

Staging directory duplicates directory structure of regular
deployment. Config generator config file is only difference now (as
far as the Makefiles are concerned).