Disable default failure alert (#1719)
* Disable Testgrid alerts by default

Modified template for testgrid to disable alerts by default on test
failures.

* Set default test group email alert

This was previously set on the dashboard tab. This may be why our alert
defaults on the test groups was not effective.

* Remove default test group email alert.

* Setting dashboard tab num_failures_to_alert to 0

See if this shuts off alerts completely for all the dashboards.

* Adding alert emails and num_failures_to_alert to dashboard tabs

Continuous, release, and nightly jobs alert if they fail
num_failures_to_alert times consecutively.