Add beta prow-tests CI flows and cleanup code
Add a copy of all continuous flows using the prow-tests:beta label and
running rarely but enough to get testing in a day across multiple
repos, also run mostly at low-use times to spread the load out.

Add comments explaining the generator better.

Use existing hash algorithm instead of the code that was here. This
will result in a one-time change to the minute that all periodic jobs
run at.