Consolidate prow-tests images (#2096)
* Consolidate prow-tests images

Now that the prow-tests image with gvm has been tested on the
beta-prow-tests flows and is working, we can consolidate down to a
single image in the jobs (which also simplifies the config generator a
little).

This first commit does not add GO_VERSION=go1.13 to the normal jobs,
to make it easy for the reviewer to validate just the images have been
changed (except for a few -go114 jobs that were missing GO_VERSION=go1.14).

Coverage is not yet figured out, so don't modify its output yet.

* Remove "go1.14" untyped string for typed struct

* Add copyright header