Ignore 'Terminating' pods when waiting for all pods running. (#2302)
Pods that are in a Terminating state don't need to be accounted for when waiting for all pods in a namespace to be up and running as the termination is something that needs to be triggered by an outside signal (i.e. a Deployment update or an explicit deletion of the pod). Those signals should be benign to the system's health.