Fix the test name and the test (#2624)
The test was not running, because name was bogus (after Test the first
letter must be capital).
And then it wouldn't pass since the strings were bogus.

Finally compare strings for equality, doing `deepequal` on strings
is a bit too much.

And then switch to use `cmp.Equal` as well.

Finally, rearrange the got/want statements.