Don't set SOURCE_DATE_EPOCH (#3074)
Since we're bumping community files and go.mod deps across our project it's
creating new nightlies with no actual code changes.

In serving this produces at least 5 PRs a day which take forever to
get through the merge pool.

This is unnecessary so this drops the use of SOURCE_DATE_EPOCH and
ko will now just zero out the timestamps.