updates rerun auth to use slugs instead of ids (#3140)
Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

Previously, only the knative org IDs were being used. This caused an
error when trying to rerun jobs in the knative-sandbox org (the team ID
was invalid, because the ID in knative did not match the ID in sandbox).

This PR updates the rerun_auth-configs to use the team slugs instead.
The teams are now identified by a combo of org and slug (which should
alleviate the problem above) and has the added benefit of being more
explicit on which team is being used (don't have to double check that
the ID is the right ID for the team).

See Slack discussion here: https://knative.slack.com/archives/C03545VN9KK/p1646761295715509